+++ src/mesh/mystring.cpp
@@ -25,7 +25,8 @@ int mystring::getstringindex(void)
 std::string mystring::getstringtonextwhitespace(void)
 {
     int indexbackup = stringindex;
-    while (stringindex < stringtoprocess.length() && stringtoprocess[stringindex] != ' ')
+    while (((size_t) stringindex < stringtoprocess.length()) &&
+           (stringtoprocess[stringindex] != ' '))
         stringindex = stringindex + 1;
     
     stringindex = stringindex + 1;
@@ -35,7 +36,8 @@ std::string mystring::getstringtonextwhitespace(void)
 
 void mystring::jumptonextwhitespace(void)
 {
-    while (stringindex < stringtoprocess.length() && stringtoprocess[stringindex] != ' ')
+    while (((size_t) stringindex < stringtoprocess.length()) &&
+           (stringtoprocess[stringindex] != ' '))
         stringindex = stringindex + 1;
     stringindex = stringindex + 1;    
 }
@@ -43,7 +45,8 @@ void mystring::jumptonextwhitespace(void)
 std::string mystring::getstringtonextcomma(void)
 {
     int indexbackup = stringindex;
-    while (stringindex < stringtoprocess.length() && stringtoprocess[stringindex] != ',')
+    while (((size_t) stringindex < stringtoprocess.length()) &&
+           (stringtoprocess[stringindex] != ','))
         stringindex = stringindex + 1;
     
     stringindex = stringindex + 1;
@@ -54,7 +57,8 @@ std::string mystring::getstringtonextcomma(void)
 std::string mystring::getstringwhileletter(void)
 {
     int indexbackup = stringindex;
-    while (stringindex < stringtoprocess.length() && std::isalpha(stringtoprocess[stringindex]))
+    while (((size_t) stringindex < stringtoprocess.length()) &&
+           std::isalpha(stringtoprocess[stringindex]))
         stringindex = stringindex + 1;
     
     return stringtoprocess.substr(indexbackup, stringindex-indexbackup);
