+++ src/shapefunction/hierarchical/hierarchicalformfunction.cpp
@@ -4,7 +4,7 @@
 int hierarchicalformfunction::gettypenumber(std::string fftypename)
 {
     int outtypenum = -1;
-    for (int i = 0; i < mytypenames.size(); i++)
+    for (size_t i = 0; i < mytypenames.size(); i++)
     {
         if (mytypenames[i] == fftypename)
         {
@@ -24,7 +24,7 @@ int hierarchicalformfunction::gettypenumber(std::string fftypename)
 
 std::string hierarchicalformfunction::gettypename(int fftypenumber)
 {
-    if (fftypenumber >= mytypenames.size())
+    if ((size_t) fftypenumber >= mytypenames.size())
     {
         std::cout << "Error in 'hierarchicalformfunction' object: unknown type number " << fftypenumber << std::endl;
         abort();
